requirements.txt
ocrmac>=1.0.0
pyperclip>=1.8.2
pynput>=1.7.6
Pillow>=10.0.0
pyobjc-core>=10.0
pyobjc-framework-Cocoa>=10.0
pyobjc-framework-ApplicationServices>=10.0
